+++ src/components/base/Base.js
@@ -589,6 +589,7 @@ export class BaseComponent {
   removeValue(index) {
     if (this.data.hasOwnProperty(this.component.key)) {
       this.data[this.component.key].splice(index, 1);
+      this.triggerChange();
     }
     this.buildRows();
   }
@@ -745,6 +746,10 @@ export class BaseComponent {
   }
 
   setInputStyles(input) {
+    if (this.labelIsHidden()) {
+      return;
+    }
+
     if (this.labelOnTheLeftOrRight(this.component.labelPosition)) {
       const totalLabelWidth = this.getLabelWidth() + this.getLabelMargin();
       input.style.width = `${100 - totalLabelWidth}%`;
@@ -757,16 +762,16 @@ export class BaseComponent {
     }
   }
 
+  labelIsHidden() {
+    return !this.component.label || this.component.hideLabel || this.options.inputsOnly;
+  }
+
   /**
    * Create the HTML element for the label of this component.
    * @param {HTMLElement} container - The containing element that will contain this label.
    */
   createLabel(container) {
-    if (
-      !this.component.label ||
-      this.component.hideLabel ||
-      this.options.inputsOnly
-    ) {
+    if (this.labelIsHidden()) {
       return;
     }
     let className = 'control-label';
@@ -905,7 +910,7 @@ export class BaseComponent {
     this.description = this.ce('div', {
       class: 'help-block'
     });
-    this.description.appendChild(this.text(this.component.description));
+    this.description.innerHTML = this.t(this.component.description);
     container.appendChild(this.description);
   }
 
@@ -1349,6 +1354,9 @@ export class BaseComponent {
   }
 
   addInputSubmitListener(input) {
+    if (!this.options.submitOnEnter) {
+      return;
+    }
     this.addEventListener(input, 'keypress', (event) => {
       let key = event.keyCode || event.which;
       if (key == 13) {
@@ -1704,16 +1712,40 @@ export class BaseComponent {
     }
 
     this._disabled = disabled;
-    // Disable all input.
-    _each(this.inputs, (input) => {
-      input.disabled = disabled;
-      if (disabled) {
-        input.setAttribute('disabled', 'disabled');
+
+    // Disable all inputs.
+    _each(this.inputs, (input) => this.setDisabled(input, disabled));
+  }
+
+  setDisabled(element, disabled) {
+    element.disabled = disabled;
+    if (disabled) {
+      element.setAttribute('disabled', 'disabled');
+    }
+    else {
+      element.removeAttribute('disabled');
+    }
+  }
+
+  setLoading(element, loading) {
+    if (element.loading === loading) {
+      return;
+    }
+
+    element.loading = loading;
+    if (!element.loader && loading) {
+      element.loader = this.ce('i', {
+        class: 'glyphicon glyphicon-refresh glyphicon-spin button-icon-right'
+      });
+    }
+    if (element.loader) {
+      if (loading) {
+        element.appendChild(element.loader);
       }
-      else {
-        input.removeAttribute('disabled');
+      else if (element.contains(element.loader)) {
+        element.removeChild(element.loader);
       }
-    });
+    }
   }
 
   selectOptions(select, tag, options, defaultValue) {