+++ src/components/textarea/TextArea.js
@@ -1,6 +1,13 @@
 import { TextFieldComponent } from '../textfield/TextField';
 import { BaseComponent } from '../base/Base';
 export class TextAreaComponent extends TextFieldComponent {
+  constructor(component, options, data) {
+    super(component, options, data);
+
+    // Never submit on enter for text areas.
+    this.options.submitOnEnter = false;
+  }
+
   wysiwygDefault() {
     return {
       theme: 'snow',