+++ src/formio.form.js
@@ -7,6 +7,7 @@ import _merge from 'lodash/merge';
 import _debounce from 'lodash/debounce';
 import _remove from 'lodash/remove';
 import _isArray from 'lodash/isArray';
+import _defaults from 'lodash/defaults';
 import _capitalize from 'lodash/capitalize';
 import EventEmitter from 'eventemitter2';
 
@@ -14,7 +15,9 @@ import EventEmitter from 'eventemitter2';
 Formio.forms = {};
 
 let getOptions = function(options) {
-  options = options || {};
+  options = _defaults(options, {
+    submitOnEnter: false
+  });
   if (!options.events) {
     options.events = new EventEmitter({
       wildcard: false,
@@ -324,11 +327,10 @@ export class FormioForm extends FormioComponents {
         var setForm = this.setForm(form);
         this.loadSubmission();
         return setForm;
-      },
-      (err) => this.formReadyReject(err)
-    ).catch(
-      (err) => this.formReadyReject(err)
-    );
+      }).catch((err) => {
+        console.warn(err);
+        this.formReadyReject(err);
+      });
   }
 
   /**
@@ -593,11 +595,10 @@ export class FormioForm extends FormioComponents {
       this.formReadyResolve();
       this.onFormBuild = null;
       this.setSubmission(this._submission);
-    },
-      (err) => this.formReadyReject(err)
-    ).catch(
-      (err) => this.formReadyReject(err)
-    );
+    }).catch((err) => {
+      console.warn(err);
+      this.formReadyReject(err);
+    });
   }
 
   /**